Refactor #88 - rename `ActionStream` to `Epic`
